refactor: HumanChatPage 사이드바 관리 훅 분리
- useHumanChatPageSideBar 훅에서 사이드바 관리가 가능함
