+++ frontend/src/pages/HumanChatPage/useHumanChatPageSIdeBar.tsx
@@ -1,0 +1,24 @@
+import { useEffect, useState } from 'react';
+
+interface useHumanChatPageSideBarParams {
+  onDisableSideBar: () => void;
+}
+export function useHumanChatPageSideBar({ onDisableSideBar }: useHumanChatPageSideBarParams) {
+  const [sideBarDisabled, setSideBarDisabled] = useState<boolean>(false);
+
+  const disableSideBar = () => {
+    onDisableSideBar();
+
+    setSideBarDisabled(true);
+  };
+
+  const enableSideBar = () => {
+    setSideBarDisabled(false);
+  };
+
+  useEffect(() => {
+    disableSideBar();
+  }, []);
+
+  return { disableSideBar, enableSideBar, sideBarDisabled };
+}
